+++ src/components/RForm/index.ts
@@ -1,0 +1,5 @@
+import RForm from './src/RForm'
+import props from './src/props'
+
+export default RForm
+export { props }