+++ src/components/RForm/src/RForm.tsx
@@ -1,0 +1,25 @@
+/**
+ *
+ * @author Ray <https://github.com/XiaoDaiGua-Ray>
+ *
+ * @date 2023-11-18
+ *
+ * @workspace ray-template
+ *
+ * @remark 今天也是元气满满撸代码的一天
+ */
+
+import { NForm } from 'naive-ui'
+
+import props from './props'
+
+export default defineComponent({
+  name: 'RForm',
+  props,
+  setup() {
+    return {}
+  },
+  render() {
+    return <NForm></NForm>
+  },
+})