+++ src/components/RForm/src/props.ts
@@ -1,0 +1,18 @@
+/**
+ *
+ * @author Ray <https://github.com/XiaoDaiGua-Ray>
+ *
+ * @date 2023-11-18
+ *
+ * @workspace ray-template
+ *
+ * @remark 今天也是元气满满撸代码的一天
+ */
+
+import { formProps } from 'naive-ui'
+
+const props = {
+  ...formProps,
+}
+
+export default props