+++ src/hooks/template/index.ts
@@ -2,9 +2,10 @@ import { useAppMenu } from './useAppMenu'
 import { useMainPage } from './useMainPage'
 import { useMenuTag } from './useMenuTag'
 import { useRootRoute } from './useRootRoute'
+import { useAppSetting } from './useAppSetting'
 
 export type { MaximizeOptions } from './useMainPage'
 export type { Target } from './useAppMenu'
 export type { CloseMenuTag } from './useMenuTag'
 
-export { useAppMenu, useMainPage, useMenuTag, useRootRoute }
+export { useAppMenu, useMainPage, useMenuTag, useRootRoute, useAppSetting }