+++ src/layout/components/SiderBar/components/SettingDrawer/index.tsx
@@ -84,20 +84,23 @@ const SettingDrawer = defineComponent({
         value: 'opacity',
       },
     ]
+    const modelSwitchReactive = reactive({
+      getMenuTagSwitch: getMenuTagSwitch.value,
+      getBreadcrumbSwitch: getBreadcrumbSwitch.value,
+      getCopyrightSwitch: getCopyrightSwitch.value,
+      getContentTransition: getContentTransition.value,
+      getWatermarkSwitch: getWatermarkSwitch.value,
+    })
 
     return {
       modelShow,
       changePrimaryColor,
       getAppTheme,
       getPrimaryColorOverride,
-      getMenuTagSwitch,
       changeSwitcher,
-      getBreadcrumbSwitch,
-      getCopyrightSwitch,
       contentTransitionOptions,
-      getContentTransition,
       updateContentTransition,
-      getWatermarkSwitch,
+      modelSwitchReactive,
     }
   },
   render() {
@@ -127,7 +130,7 @@ const SettingDrawer = defineComponent({
               {$t('headerSettingOptions.ContentTransition')}
             </NDivider>
             <NSelect
-              v-model:value={this.getContentTransition}
+              v-model:value={this.modelSwitchReactive.getContentTransition}
               options={this.contentTransitionOptions}
               onUpdateValue={(value) => {
                 this.updateContentTransition(value)
@@ -139,7 +142,7 @@ const SettingDrawer = defineComponent({
             <NDescriptions labelPlacement="left" column={1}>
               <NDescriptionsItem label="多标签">
                 <NSwitch
-                  v-model:value={this.getMenuTagSwitch}
+                  v-model:value={this.modelSwitchReactive.getMenuTagSwitch}
                   onUpdateValue={(bool: boolean) =>
                     this.changeSwitcher(bool, 'menuTagSwitch')
                   }
@@ -147,7 +150,7 @@ const SettingDrawer = defineComponent({
               </NDescriptionsItem>
               <NDescriptionsItem label="面包屑">
                 <NSwitch
-                  v-model:value={this.getBreadcrumbSwitch}
+                  v-model:value={this.modelSwitchReactive.getBreadcrumbSwitch}
                   onUpdateValue={(bool: boolean) =>
                     this.changeSwitcher(bool, 'breadcrumbSwitch')
                   }
@@ -155,7 +158,7 @@ const SettingDrawer = defineComponent({
               </NDescriptionsItem>
               <NDescriptionsItem label="水印">
                 <NSwitch
-                  v-model:value={this.getWatermarkSwitch}
+                  v-model:value={this.modelSwitchReactive.getWatermarkSwitch}
                   onUpdateValue={(bool: boolean) =>
                     this.changeSwitcher(bool, 'watermarkSwitch')
                   }
@@ -163,7 +166,7 @@ const SettingDrawer = defineComponent({
               </NDescriptionsItem>
               <NDescriptionsItem label="版权信息">
                 <NSwitch
-                  v-model:value={this.getCopyrightSwitch}
+                  v-model:value={this.modelSwitchReactive.getCopyrightSwitch}
                   onUpdateValue={(bool: boolean) =>
                     this.changeSwitcher(bool, 'copyrightSwitch')
                   }