+++ src/utils/vue/index.ts
@@ -1,5 +1,7 @@
-export { call } from './call'
-export { unrefElement } from './unrefElement'
-export { renderNode } from './renderNode'
-export { effectDispose } from './effectDispose'
-export { watchEffectWithTarget } from './watchEffectWithTarget'
+import { call } from './call'
+import { unrefElement } from './unrefElement'
+import { renderNode } from './renderNode'
+import { effectDispose } from './effectDispose'
+import { watchEffectWithTarget } from './watchEffectWithTarget'
+
+export { call, unrefElement, renderNode, effectDispose, watchEffectWithTarget }